Fizx lint and table_manager.go was getting a bit big.
