+++ pkg/chunk/table_client.go
@@ -1,0 +1,42 @@
+package chunk
+
+import "golang.org/x/net/context"
+
+// TableClient is a client for telling Dynamo what to do with tables.
+type TableClient interface {
+	ListTables(ctx context.Context) ([]string, error)
+	CreateTable(ctx context.Context, desc TableDesc) error
+	DescribeTable(ctx context.Context, name string) (desc TableDesc, status string, err error)
+	UpdateTable(ctx context.Context, desc TableDesc) error
+}
+
+// TableDesc describes a table.
+type TableDesc struct {
+	Name             string
+	ProvisionedRead  int64
+	ProvisionedWrite int64
+	Tags             Tags
+}
+
+// Equals returns true if other matches desc.
+func (desc TableDesc) Equals(other TableDesc) bool {
+	if desc.ProvisionedRead != other.ProvisionedRead {
+		return false
+	}
+
+	if desc.ProvisionedWrite != other.ProvisionedWrite {
+		return false
+	}
+
+	if !desc.Tags.Equals(other.Tags) {
+		return false
+	}
+
+	return true
+}
+
+type byName []TableDesc
+
+func (a byName) Len() int           { return len(a) }
+func (a byName) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+func (a byName) Less(i, j int) bool { return a[i].Name < a[j].Name }
